backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
google-generativeai>=0.8.0
